implemented "direction:vertical" support, also for symbols
git-svn-id: svn://svn.code.sf.net/p/magicseteditor/code/trunk@269 0fc631ac-6414-0410-93d0-97cfa31319b6

+++ src/gui/print_window.cpp
@@ -120,7 +120,7 @@ class CardsPrintout : wxPrintout {
 	PageLayout layout;
 	
 	/// Draw a card, that is card_nr on this page, find the postion by asking the layout
-	void drawCard(DC& dc, const CardP& card, UInt card_nr);
+	void drawCard(DC& dc, const CardP& card, size_t card_nr);
 	/// Draw a card at the specified coordinates
 	void drawCard(DC& dc, const CardP& card, double x, double y, int rotation = 0);
 };
